chore(FDS-438): [Boundaries] - move Boundaries into atoms folder
