+++ packages/cascara/src/atoms/index.js
@@ -2,6 +2,7 @@
 // Everything exported here is being exported from Cascara.
 // Let's keep these sorted alphabetically based on their source directory.
 
+export { default as Boundaries } from './Boundaries';
 export { default as Button } from './Button';
 export { default as Flex } from './Flex';
 export { default as Image } from './Image';
