+++ packages/cascara/src/system-components/index.js
@@ -3,5 +3,4 @@
 // Let's keep these sorted alphabetically based on their source directory.
 
 export { default as AreaPlaceholder } from './AreaPlaceholder';
-export { default as Boundaries } from './Boundaries';
 export { default as JsonPlaceholder } from './JsonPlaceholder';
