+++ packages/cascara/src/ui/FormOld/FormOld.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import pt from 'prop-types';
 // import styles from './FormOld.module.scss';
 
-import { Boundaries } from '../../system-components';
+import { Boundaries } from '../../atoms';
 import FormProvider from './context/FormProvider';
 import ModuleError from '../../modules/ModuleError';
