+++ packages/cascara/src/ui/FormOld/modules/FormRow/FormRow.js
@@ -2,7 +2,7 @@ import React from 'react';
 import pt from 'prop-types';
 
 import styles from '../../FormOld.module.scss';
-import { Boundaries } from '../../../../system-components';
+import { Boundaries } from '../../../../atoms';
 
 const propTypes = {
   children: pt.oneOf([pt.node, pt.arrayOf(pt.node)]),
