+++ packages/cascara/src/ui/Section/Section.js
@@ -3,7 +3,7 @@ import pt from 'prop-types';
 import { Role } from 'reakit';
 import classNames from 'classnames/bind';
 
-import Boundaries from '../../system-components/Boundaries';
+import Boundaries from '../../atoms/Boundaries';
 import Title from '../Title';
 import { PROP_TYPES as TitlePT } from '../Title/__globals';
