+++ packages/cascara/src/ui/TableOld/TableBody.js
@@ -1,7 +1,7 @@
 import React, { useContext, useMemo } from 'react';
 import styles from './TableOld.module.scss';
 
-import { Boundaries } from '../../system-components';
+import { Boundaries } from '../../atoms';
 import { ModuleContext } from '../../modules/context';
 import TableRow from './TableRow';
